Improve DownloadInfo in EhDB

+++ app/src/main/java/com/hippo/ehviewer/client/EhCacheKeyFactory.java
@@ -18,23 +18,23 @@ package com.hippo.ehviewer.client;
 
 public class EhCacheKeyFactory {
 
-    public static String getThumbKey(int gid) {
+    public static String getThumbKey(long gid) {
         return "thumb:" + gid;
     }
 
-    public static String getNormalPreviewKey(int gid, int index) {
+    public static String getNormalPreviewKey(long gid, int index) {
         return "preview:normal:" + gid + ":" + index;
     }
 
-    public static String getLargePreviewKey(int gid, int index) {
+    public static String getLargePreviewKey(long gid, int index) {
         return "preview:large:" + gid + ":" + index;
     }
 
-    public static String getLargePreviewSetKey(int gid, int index) {
+    public static String getLargePreviewSetKey(long gid, int index) {
         return "large_preview_set:" + gid + ":" + index;
     }
 
-    public static String getImageKey(int gid, int index) {
+    public static String getImageKey(long gid, int index) {
         return "image:" + gid + ":" + index;
     }
 }
