+++ app/src/main/java/com/hippo/ehviewer/ui/scene/GalleryInfoScene.java
@@ -81,7 +81,7 @@ public final class GalleryInfoScene extends ToolbarScene implements EasyRecycler
         mKeys.add(resources.getString(R.string.header_key));
         mValues.add(resources.getString(R.string.header_value));
         mKeys.add(resources.getString(R.string.key_gid));
-        mValues.add(Integer.toString(gd.gid));
+        mValues.add(Long.toString(gd.gid));
         mKeys.add(resources.getString(R.string.key_token));
         mValues.add(gd.token);
         mKeys.add(resources.getString(R.string.key_url));
